build.sh: Configurable binary name
The name of the build output was enforced to project name. This patch
adds a configurable name for the build output dictated by BIN_NAME. The
the output will default to $PROJECT_NAME$EXT if BIN_NAME is left unset.
Note that `BUILD_FLAGS="-o $BIN_NAME"` must be used if the
`basename $CMD_PATH` != BIN_NAME.