build.sh: Warn if OUTPUT is set to non-existing file
OUTPUT can be altered through various configurations and environment
variables. This can cause OUTPUT to contain a name which is not pointing
to a file.

This patch adds a check and a warning message if OUTPUT contains a name
for a non-existing file before returning it. This should improve
debugging.